Set config::MAX_SERVICE_CLIENTS based on clients-$N feature
The number of supported clients is configurable through a crate feature,
but the feature is only honored in the Interchange handling in src/pipe.rs.
The config::MAX_SERVICE_CLIENTS type, which is used to set the size of the
"endpoints" Vec<> in src/service.rs, was still hardcoded to consts::U5.