Enforce namespace in FilesystemClient
This patch adds a check to FilesystemClient that makes sure that clients
cannot read out files that belong to other clients or that are not in
the data directory for the client.  To keep the check simple, we
completely forbid `..` in paths.

This is a breaking change as the Error enum is extended.

Fixes https://github.com/trussed-dev/trussed/issues/65