Remove serde trait implementations for requests and replies
Implementing Serialize and Deserialize for the request and reply structs
leaks implementation details, especially when using serde_indexed.  This
patch removes these implementations for the request and reply structs
and also for some other types that presumably only had them because they
were used in these structs and that are not serialized or deserialized
anywhere in the Trussed ecosystem.

To make it easier to store encrypted data – where previously
reply::Encrypt could be used direcly – this patch adds an EncryptedData
struct that implemente Serialize and Deserialize.

Fixes: https://github.com/trussed-dev/trussed/issues/183