core: Make more types non-exhaustive
This patch marks more types as non-exhaustive that could be extended in
the future and that don’t need to be matched exhaustively:
- KeySerialization
- SignatureSerialization
- consent::Error

The only exhaustive types that could be realistically be extended in
the future are now the request and reply structs.  But marking these as
non-exhaustive would make it very complex for the backends to implement
these syscalls.  We can try to find a solution for that when we think
about alternative syscall implementations, e. g. using a builder
pattern.