core: Flatten public module structure
This patch makes some modules of the core crate private and re-exports
their contents from the crate root.  This has two benefits:  It makes
it easier to read the crate documentation because the most important
types are directly visible from the root.  And it makes the imports for
the most important types less verbose.

See also: https://github.com/trussed-dev/trussed/issues/155