core: Add helpers for checking mechanisms at compile time
To be able to compare mechanism variants in const contexts, we use
add a const_eq function.  And to be able to indicate which mechanism
is potentially missing, we add a const panic method that includes the
mechanism variant in the panic message.

To reduce the boilerplate code, we use a macro to generate the enum.