Remove Syscall implementations for Service
Trussed is intended to be executed in a separate task.  Implementing
Syscall for Service directly should not be necessary and makes the types
more complex.  Instead, hardware runners should use interrupts and
software runners should use channels (like in trussed::virt) to
communicate with the service.  If necessary, a wrapper struct can be
used for the old behavior.