Simplify Store trait
This patch simplifies the Store trait by removing the Fs wrapper struct,
the Storage types, the static lifetime and the Copy requirement and
replaces it with a reference to a DynFilesystem.

This gives runners more options on how to implement the store.

As the Store trait can now easily be implemented in a safe way, this
patch also removes the unsafe keyword from the trait definition.