Use single source of truth for implemented mechanisms
This patch removes the manual mechanism enumeration in
IMPLEMENTED_MECHANISMS and changes the rpc_trait! macro to also generate
this constant so that it acts as a single source of truth.  As a side
effect, the IMPLEMENTED_MECHANISMS constant is moved from types to
service and only available if the crypto-client feature is enabled.