Make mechanisms module private
This patch makes the mechanisms module private.  The mechanism
implementation can still be accessed using the Mechanism enum.  This
leads to a clenaer public API and triggers a compiler warning if the
mapping between the Mechanism enum and the mechanism implementations is
wrong (as the implementation would then be unused).